reworked image API

+++ src/ui.h
@@ -23,7 +23,7 @@ public:
     std::string selected_output_image;
 
     int render_scale_index = 0;
-    ivec3 color_under_cursor = ivec3(0);
+    vec3 color_under_cursor = vec3(0);
     int max_ray_depth = 5;
     int frame_samples = 4;
     bool direct_lighting_enabled;
